+++ src/TensorMixedStates.jl
@@ -9,6 +9,7 @@ import LinearAlgebra: dot
 import ITensorMPS: maxlinkdim, apply, state, expect, normalize, measure!, checkdone!, tdvp, dmrg, sample
 
 using ITensors, ITensorMPS, Printf, Dates, JSON, Random, LinearAlgebra, HDF5
+import Logging
 
 # Core
 include("Operators.jl")